+++ src/commands/fun/urban.ts
@@ -1,27 +1,31 @@
 import {Command, NamedCommand} from "../../core";
 import {MessageEmbed} from "discord.js";
-// Anycasting Alert
-const urban = require("relevant-urban");
+import urban from "relevant-urban";
 
 export default new NamedCommand({
     description: "Gives you a definition of the inputted word.",
-    async run({message, channel, guild, author, member, client, args}) {
-        if (!args[0]) {
-            channel.send("Please input a word.");
+    run: "Please input a word.",
+    any: new Command({
+        async run({message, channel, guild, author, member, client, args}) {
+            // [Bug Fix]: Use encodeURIComponent() when emojis are used: "TypeError [ERR_UNESCAPED_CHARACTERS]: Request path contains unescaped characters"
+            urban(encodeURIComponent(args.join(" ")))
+                .then((res) => {
+                    const embed = new MessageEmbed()
+                        .setColor(0x1d2439)
+                        .setTitle(res.word)
+                        .setURL(res.urbanURL)
+                        .setDescription(`**Definition:**\n*${res.definition}*\n\n**Example:**\n*${res.example}*`)
+                        // [Bug Fix] When an embed field is empty (if the author field is missing, like the top entry for "british"): "RangeError [EMBED_FIELD_VALUE]: MessageEmbed field values may not be empty."
+                        .addField("Author", res.author || "N/A", true)
+                        .addField("Rating", `**\`Upvotes: ${res.thumbsUp} | Downvotes: ${res.thumbsDown}\`**`);
+                    if (res.tags && res.tags.length > 0 && res.tags.join(" ").length < 1024)
+                        embed.addField("Tags", res.tags.join(", "), true);
+
+                    channel.send(embed);
+                })
+                .catch(() => {
+                    channel.send("Sorry, that word was not found.");
+                });
         }
-        const res = await urban(args.join(" ")).catch((e: Error) => {
-            return channel.send("Sorry, that word was not found.");
-        });
-        const embed = new MessageEmbed()
-            .setColor(0x1d2439)
-            .setTitle(res.word)
-            .setURL(res.urbanURL)
-            .setDescription(`**Definition:**\n*${res.definition}*\n\n**Example:**\n*${res.example}*`)
-            .addField("Author", res.author, true)
-            .addField("Rating", `**\`Upvotes: ${res.thumbsUp} | Downvotes: ${res.thumbsDown}\`**`);
-        if (res.tags.length > 0 && res.tags.join(" ").length < 1024) {
-            embed.addField("Tags", res.tags.join(", "), true);
-        }
-        channel.send(embed);
-    }
+    })
 });