+++ src/commands/utility/emote.ts
@@ -2,8 +2,9 @@ import {Command, NamedCommand} from "../../core";
 import {processEmoteQueryFormatted} from "./modules/emote-utils";
 
 export default new NamedCommand({
-    description: "Send the specified emote.",
-    run: "Please provide a command name.",
+    description:
+        "Send the specified emote list. Enter + to move an emote list to the next line, - to add a space, and _ to add a zero-width space.",
+    run: "Please provide a list of emotes.",
     any: new Command({
         description: "The emote(s) to send.",
         usage: "<emotes...>",